Fixed issue with partial_load_file.txt take 2
Trying to merge these

+++ CPC_TSI.py
@@ -121,7 +121,8 @@ def remove_previous_output(filetype, reload_from_source):
         # Delete files and return
         for file in filelist:
             os.remove(file)
-        os.remove('files_loaded.txt')
+        if os.path.isfile('files_loaded.txt'):
+            os.remove('files_loaded.txt')
     return
 
 def save_to_hdf(data, output_h5_filename, output_file_frequency):    
@@ -279,15 +280,14 @@ def read_cpc_csv(read_filename, output_filename_base, output_file_frequency, Inp
     numsamples = lastline.split(",")[0]
     
     # Read where the import has gotten up to previously:
+    first_load = True
     if os.path.isfile('partial_files_loaded.txt'):
         with open('partial_files_loaded.txt','r') as f:
             last_loaded = f.readlines()
             last_loaded = [x.strip() for x in last_loaded]
             last_loaded_file = last_loaded[0]
             last_loaded_sample = int(last_loaded[1])
-            files_previously_loaded = True
-    else:
-        files_previously_loaded = False
+            first_load = False
     
     for chunk in df:
         # Extract initial timestamp for each sample (i.e. each row)
@@ -299,14 +299,14 @@ def read_cpc_csv(read_filename, output_filename_base, output_file_frequency, Inp
         
         data = pd.DataFrame(columns = {'Timestamp', 'Concentration'})
         for rowidx in range(0,len(chunk)):
-            if files_previously_loaded:
+            if not first_load:
                 if (chunk['Sample #'][rowidx] <= last_loaded_sample) and (read_filename == last_loaded_file):
                     continue
             # Create timestamp and extract concentration for each sample in chunk
-            timestamp = [chunk['sample_timestamp'][rowidx] + pd.Timedelta(seconds=x) for x in range(0, chunk['Sample Length'][rowidx])]
-            conc = chunk.iloc[rowidx,12:(12+chunk['Sample Length'][rowidx])]
+            timestamp = [chunk['sample_timestamp'][rowidx]+pd.Timedelta(seconds=x) for x in range(0,chunk['Sample Length'][rowidx])]            
+            conc = chunk.loc[rowidx][12:(12+chunk['Sample Length'][rowidx])]
             
-            print('Formatting sample ' + str(chunk['Sample #'].iloc[rowidx]) 
+            print('Formatting sample ' + str(chunk['Sample #'].loc[rowidx]) 
                   + ' of ' + numsamples + ' from file ' + read_filename)
             
             # Format data as dataframe
